+++ src/Application/ReadModel/StockLevel.php
@@ -6,35 +6,35 @@ namespace Application\ReadModel;
 final class StockLevel
 {
     /**
-     * @var float
+     * @var int
      */
     private $quantityInStock;
 
-    private function __construct(float $quantityInStock)
+    private function __construct(int $quantityInStock)
     {
         $this->quantityInStock = $quantityInStock;
     }
 
     public static function initial(): StockLevel
     {
-        return new StockLevel(0.0);
+        return new StockLevel(0);
     }
 
-    public function add(float $quantity): StockLevel
+    public function add(int $quantity): StockLevel
     {
         return new self(
             $this->quantityInStock + $quantity
         );
     }
 
-    public function subtract(float $quantity): StockLevel
+    public function subtract(int $quantity): StockLevel
     {
         return new self(
             $this->quantityInStock - $quantity
         );
     }
 
-    public function asFloat(): float
+    public function asInt(): int
     {
         return $this->quantityInStock;
     }
